GCC updates from todays work
git-svn-id: https://fairport.svn.codeplex.com/svn/fairport/trunk@38393 c102de6d-fbcd-4082-88ef-c2bab1378b82

+++ fairport/ndb/node.h
@@ -563,7 +563,7 @@ inline std::streampos fairport::node_stream_device::seek(boost::iostreams::strea
 
     if(m_pos < 0)
         m_pos = 0;
-    else if(m_pos > m_pnode->size())
+    else if(static_cast<size_t>(m_pos) > m_pnode->size())
         m_pos = m_pnode->size();
 
     return m_pos;
